requirements.txt
firecrawl
pydantic
python-dotenv
fastapi
python-multipart
uvicorn
gunicorn
